Heavily improved efficiency of conjugated verbs
We no longer loop over the entire list of conjugated word lists, and then checking if our verb is inside of this smaller list, for each verb. Instead, the Conjugated verb list is now a dictionary. Each of the conjugated verbs are used as a dictionary to point to the same Verb object, which stores the small list of related conjugated verbs.