Simplification and improvements of Adjective to Adverb additions
ADJECTIVE_TO_ADVERB default dictionary simplified to remove cases where the key and value are the same, as these don't end up doing anything for us.
Pertainyms is now a set to avoid unnecessary duplicates.
Now each of the pertainyms are added in the dict, rather than just the first.